sundial class migrated to ts
